+++ src/mfclib/config.py
@@ -1,7 +1,9 @@
 import warnings
 from typing import Any, Dict
 
+import deprecation
 import pint
+from .version import __version__
 
 UNIT_REGISTRY_KEY = 'UNIT_REGISTRY'
 BALANCE_SPECIES_KEY = 'BALANCE_SPECIES_INDICATOR'
@@ -48,14 +50,27 @@ def balanceSpeciesIndicator():
     return indicator
 
 
-def unitRegistry() -> None | pint.UnitRegistry:
-    return _get_config_value(UNIT_REGISTRY_KEY)
+def unit_registry():
+    if ureg := _get_config_value(UNIT_REGISTRY_KEY):
+        return ureg
+    else:
+        return pint.get_application_registry()
 
 
 def sourceGasesFile():
     return _get_config_value(SOURCE_GASES_FILE_KEY)
 
 
+@deprecation.deprecated(
+    deprecated_in='0.2.3',
+    removed_in='1.0',
+    current_version=__version__,
+    details=(
+        'Pint now already includes "%" and "ppm" as units,'
+        ' so an explicit registration is not necessary.'
+        ' "frac" and "ppb" will be removed as units entirely.'
+    ),
+)
 def register_units(registry: pint.UnitRegistry):
     """
     Registers specialized dimensionless units in the provided unit registry.